test: split view, navigation, gallery and status bar out of card.test.ts
card.test.ts was still 78K after the auto_cycle/auto_zoom split. Four
more behaviour groupings move out:

  card.gallery.test.ts     strip, panel, source resolution, retries
  card.view.test.ts        viewBox fitting, zoom levels, drag-to-pan, Now highlight
  card.navigation.test.ts  hour/day/month steps, replay, button labels, date readout
  card.status-bar.test.ts  status bar layout, version badge, debug overlay

card.test.ts keeps what doesn't belong to any of them: config parsing,
height/theme appearance, the hass setter and location resolution, and the
lifecycle null guards. 1935 lines down to 372.

The gallery block carried its own createAndMount that forces
gallery.mode "both" and shadowed the outer helper of the same name. In
its own file that shadowing became a silent difference from the shared
helper, so it is now mountWithGallery.

No source change, no test added or removed: 801 before, 801 after.
